requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
inflection==0.5.1
install==1.3.5
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
minio==7.2.5
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
python-environ==0.4.54
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rest-framework-simplejwt==0.0.2
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
